refactor: consolidate network icon components
