+++ src/app/components/FilterNetwork/NetworkOptions.tsx
@@ -3,16 +3,14 @@ import React from "react";
 import { FilterOption } from "./FilterNetwork.contracts";
 import { Badge } from "@/app/components/Badge";
 import { Circle } from "@/app/components/Circle";
-import { Tooltip } from "@/app/components/Tooltip";
-import { networkDisplayName } from "@/utils/network-utils";
-import { NetworkIconContent } from "@/app/components/SelectNetworkDropdown/SelectNetworkDropdown.blocks";
+import { NetworkIcon } from "@/domains/network/components/NetworkIcon";
 
 export const NetworkOption = ({ network, isSelected, onClick }: FilterOption) => {
 	const renderOption = () => {
 		if (isSelected) {
 			return (
-				<Circle size="lg" className="relative border-theme-primary-500 text-theme-primary-500">
-					<NetworkIconContent network={network} />
+				<Circle size="lg" className="relative border-theme-primary-500">
+					<NetworkIcon iconClassName="text-theme-primary-500" network={network} isCompact />
 					<Badge
 						className="border-transparent bg-theme-primary-500 text-theme-primary-100"
 						icon="CheckmarkSmall"
@@ -24,9 +22,9 @@ export const NetworkOption = ({ network, isSelected, onClick }: FilterOption) =>
 		return (
 			<Circle
 				size="lg"
-				className="relative border-theme-secondary-300 text-theme-secondary-300 dark:border-theme-secondary-800"
+				className="relative border-theme-secondary-300 dark:border-theme-secondary-800"
 			>
-				<NetworkIconContent network={network} />
+				<NetworkIcon iconClassName="text-theme-secondary-300" network={network} isCompact />
 				<Badge className="border-theme-secondary-300 dark:border-theme-secondary-800" />
 			</Circle>
 		);
@@ -38,7 +36,7 @@ export const NetworkOption = ({ network, isSelected, onClick }: FilterOption) =>
 			data-testid={`NetworkOption__${network.id()}`}
 			onClick={onClick}
 		>
-			<Tooltip content={networkDisplayName(network)}>{renderOption()}</Tooltip>
+			{renderOption()}
 		</li>
 	);
 };
