+++ src/app/components/SelectNetworkDropdown/SelectNetworkDropdown.blocks.tsx
@@ -1,14 +1,15 @@
 import React from "react";
 import { Networks } from "@ardenthq/sdk";
-import cn from "classnames";
 import { useTranslation } from "react-i18next";
 import { Tooltip } from "@/app/components/Tooltip";
 import { Icon } from "@/app/components/Icon";
 import { isCustomNetwork, networkDisplayName } from "@/utils/network-utils";
 import { Size } from "@/types";
+import { NetworkIcon } from "@/domains/network/components/NetworkIcon";
 
 export const NetworkTestnetCodeIcon = () => {
 	const { t } = useTranslation();
+
 	return (
 		<Tooltip content={t("COMMON.TEST_NETWORK")}>
 			<span>
@@ -18,38 +19,6 @@ export const NetworkTestnetCodeIcon = () => {
 	);
 };
 
-export const NetworkIconContent = ({ network, iconSize }: { network: Networks.Network; iconSize?: Size }) => {
-	if (isCustomNetwork(network)) {
-		return <>{networkDisplayName(network).slice(0, 2).toUpperCase()}</>;
-	}
-
-	return <Icon data-testid="NetworkIcon__icon" name={network.ticker()} size={iconSize} />;
-};
-
-export const NetworkIcon = ({
-	network,
-	className,
-	iconSize,
-}: {
-	network: Networks.Network;
-	className?: string;
-	iconSize?: Size;
-}) => (
-	<div
-		className={cn(
-			"flex items-center justify-center rounded-xl p-1",
-			{
-				"text-theme-primary-600": network.isLive(),
-				"text-theme-secondary-700": !network.isLive(),
-			},
-			className,
-		)}
-		aria-label={networkDisplayName(network)}
-	>
-		<NetworkIconContent network={network} iconSize={iconSize} />
-	</div>
-);
-
 export const NetworkOptionLabel = ({ network }: { network?: Networks.Network }) => {
 	if (!network) {
 		return <></>;
@@ -57,7 +26,7 @@ export const NetworkOptionLabel = ({ network }: { network?: Networks.Network }) 
 
 	return (
 		<div className="flex items-center space-x-3">
-			<NetworkIcon network={network} />
+			<NetworkIcon network={network} showTooltip={false} isCompact />
 			<div className="flex-grow">{networkDisplayName(network)}</div>
 			{network.isTest() && <NetworkTestnetCodeIcon />}
 		</div>
