+++ src/domains/network/components/NetworkOptions/NetworkOptions.tsx
@@ -4,10 +4,8 @@ import React, { memo } from "react";
 import { wrapperClasses, optionClasses } from "./NetworkOptions.styles";
 import { NetworkOptionsProperties, NetworkOptionProperties } from "./NetworkOptions.contracts";
 import { networkDisplayName } from "@/utils/network-utils";
-import {
-	NetworkTestnetCodeIcon,
-	NetworkIcon,
-} from "@/app/components/SelectNetworkDropdown/SelectNetworkDropdown.blocks";
+import { NetworkTestnetCodeIcon } from "@/app/components/SelectNetworkDropdown/SelectNetworkDropdown.blocks";
+import { NetworkIcon } from "@/domains/network/components/NetworkIcon";
 
 export const NetworkOption = memo(
 	({ disabled, network, iconSize = "lg", isSelected, onSelect, onDeselect }: NetworkOptionProperties) => {
@@ -32,7 +30,7 @@ export const NetworkOption = memo(
 					data-testid={`NetworkOption-${network.coin()}-${network.id()}`}
 				>
 					<div className="flex items-center space-x-4">
-						<NetworkIcon network={network} iconSize={iconSize} />
+						<NetworkIcon network={network} iconSize={iconSize} showTooltip={false} isCompact />
 
 						<div className="font-semibold text-theme-secondary-700 dark:text-theme-secondary-200">
 							{networkDisplayName(network)}
